TASK B (IN PROGRESS): Adds bean level and database level constraints and validations

+++ backend/playlytics/src/main/java/com/mparker/playlytics/composite_ids/ConfirmedConnectionId.java
@@ -12,9 +12,9 @@ public class ConfirmedConnectionId implements Serializable {
     // Database Columns
 
     @Column(name = "registered_player_A_id")
-    private int registeredPlayerAId;
+    private long registeredPlayerAId;
 
     @Column(name = "registered_player_B_id")
-    private int registeredPlayerBId;
+    private long registeredPlayerBId;
 
 }
